getCurrentTemperature conversion to float and fixex

+++ src/main/sensors/temperature.h
@@ -27,5 +27,5 @@ typedef enum {
 
 // Temperature is returned in degC*10
 int16_t getTemperature(tempSensor_e sensor);
-int16_t getCurrentTemperature(void);
+float getCurrentTemperature(void);
 void temperatureUpdate(void);
